viz: gather constants & buffer into Rms structure
finishing the modularization
